requirements.txt
fastapi==0.119.0
uvicorn==0.37.0
python-multipart==0.0.20
python-dotenv==1.1.1
langgraph==1.0.8
google-genai==1.62.0
groq==1.0.0
tavily-python==1.1.0
llama-parse==0.6.94
pydantic==2.11.7
nest-asyncio==1.6.0